Move Google sign-in button to main index
